переделал структуру проекта, у комментов собственное приложение
