+++ README.md
@@ -31,6 +31,7 @@
     + текстовый комментарий отображается в карточке пользователя
     + оценки отображаются в карточке пользователя в качестве средней оценки
     + в профиле отображаются купленные товары, на которую ещё не оставлен отзыв
+    - Изменить схему для обработки form_invalid в add_comment
     - Добавить пермишны на удаление и редактирование комментария
     - Добавить валидаторы на оценки
     - Добавить фильтр матерных слов
