+++ coffeeshop/constants.py
@@ -1,0 +1,2 @@
+NAME_LENGTH_LIMIT = 64
+ADMIN_LENGTH_LIMIT = 20
