+++ goods/views.py
@@ -1,11 +1,11 @@
 from django.http import Http404
 from django.db.models import Avg
 from django.shortcuts import get_object_or_404
-from django.urls import reverse
-from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
 
-from goods.forms import CommentForm
-from goods.models import Category, Comment, Product
+from django.views.generic import ListView, DetailView
+
+from comments.forms import CommentForm
+from goods.models import Category, Product
 
 
 class CatalogView(ListView):
@@ -49,42 +49,3 @@ class ProductView(DetailView):
         context["comment_form"] = CommentForm()  
 
         return context
-
-# Представления комментариев
-# (Необходимо придумать структуру, чтобы обрабатывать
-# ошибки при вводе коментариев)
-
-class BaseCommentView:
-    model = Comment
-
-    def get_success_url(self):
-        product = Product.objects.get(pk = self.object.product.id)
-        return reverse(
-            "catalog:product",
-            kwargs = {"product_slug": product.slug},
-        )
-
-
-class AddCommentView(BaseCommentView, CreateView):  
-    form_class = CommentForm
-    template_name = "goods/product.html"
-
-    def form_valid(self, form):  
-        form.instance.author = self.request.user  
-        form.instance.product = Product.objects.get(pk=self.kwargs.get("pk"))  
-        return super().form_valid(form)
-    
-    def form_invalid(self, form):
-        response = super().form_invalid(form)
-        return response    
-
-
- # Добавить пермишн на проверку - только автор или админ
-class EditCommentView(BaseCommentView, UpdateView): 
-    form_class = CommentForm  
-    template_name = "comments/comment_edit.html"  
-
-
-# Добавить пермишн на проверку - только автор или админ
-class DeleteCommentView(BaseCommentView, DeleteView):  
-    template_name = "comments/comment_delete.html"
